requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.3.2